+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 
 #### TWEAKS
 - `[UI/UX]` Normalize the size of all buttons
-- `[Lidarr/Releases]` Use the new expandable table card styling
-- `[Tautulli/Logs]` Use the new expandable table card styling
+- `[Lidarr/Releases]` Use the newer expandable table card styling
+- `[SABnzbd/History]` Use the newer expandable table card styling
+- `[Tautulli/Logs]` Use the newer expandable table card styling
 
 #### FIXES
 - `[Tautulli/Activity]` Activity header included some incorrectly localized strings